hidpp10: fix 0x63 sensor resolution
This is a short or long request, depending on the hardware. The short request
may return one of two data sets but we don't know which one it will be without
a hw database. The resolution bits are identical in both though.

For the short/long - the G500s errors out on a short request, let's assume
this is true for other devices too and poke the device with the short request
first, then with the long request. If one succeeds, we continue from there.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>