driver-hidpp10: hook up writing the resolution
Long term we want to get rid of this, it's better to write the resolution in
the profile and update the profile than just toggling the current resolution
(which requires libratbag changes elsewhere to find it again if it changes
from the ones in the current profile).

But for now, it at least allow testing of the 0x63 request.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>