Use breathe for auto-doxygen
This is worse than our custom XSL style sheet because it doesn't give us
decent references. With the XSL we can use cpp enums and link to them, I
couldn't find a way to make breathe spit out proper references. So we need to
include them manually (and add :nolink:) but then we still cannot link to the
actual enum values.

And we still need to run doxygen manually, so even that doesn't give us much