hidpp10: assert that our error code is negative
This is to shut up scan-build which seems to think that we may return a
positive number here on error. This way it can assume a negative or zero
number in the caller and stops complaining about

../../../src/hidpp10.c:1171:11: warning: Dereference of null pointer
                switch (m->any.type) {

Only hidpp10_onboard_profiles_read_macro() is needed to shut this up, but
let's put it in for two other functions too.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>