hidpp10: shut up scan build
For some reason scan-build seems to think that page can be uninitialized on
the call to long_set_bit() a few lines above. This later triggers an error on
the ffs() call here:

../../../src/hidpp10.c:1464:10: warning: 1st function call argument is an uninitialized value
                page = ffsl(pages) - 1;

AFAICT this is rubbish. All memory is zeroed so page should be either 0 or a
valid value. The scan-build logs have no indication of *why* this value is
assumed uninintialized so let's go with ifdefing out this false positive.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>