hidpp20: don't decrease the index on a JUMP macro entry
We never increase it for those entries, so if our macros is JUMP and something
else, we decrease index to -1, then increase it to 0 on the 'something else'
bit. We return 0 which is treated as an error but we have a macro allocated
then, thus leaking the memory.

The index-- instruction was presumably done with the assumption that the index
increases with every command - but that's only done in the else condition. So
let's drop the decrease.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>