ratbagd: make the profile vtable dynamic
The DBus API shouldn't export properties that do not exist on the device. The
goal here is that where a property "foo" is available, that property is also
writable by the user. Given that DBus itself is dynamic this gives us a
tristate for everything, yes/no/does-not-exist.

However.... the sd-bus API is not really designed for this, so we have to work
around this a bit. We build a dynamic vtable instead of the previous fallback
one but have to use a macro because we can only do it one-by-one.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>